personal_os/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.2.0
onnxruntime-cpu>=1.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
litellm>=1.0.0
GitPython>=3.1.0
PyYAML>=6.0
regex>=2024.0.0
filelock>=3.0.0
